+++ pygnssutils/__init__.py
@@ -14,6 +14,7 @@ from pygnssutils.exceptions import (
 from pygnssutils.gnssdump import GNSSStreamer
 from pygnssutils.gnssserver import GNSSSocketServer
 from pygnssutils.gnssntripclient import GNSSNTRIPClient
+from pygnssutils.gnssmqttclient import GNSSMQTTClient
 from pygnssutils.ubxsetrate import UBXSetRate
 from pygnssutils.ubxload import UBXLoader
 from pygnssutils.ubxsave import UBXSaver